Move References.kt to connection package in common
