+++ wearos/src/main/java/com/boswelja/devicemanager/batterysync/PhoneBatteryUpdateReceiver.kt
@@ -15,10 +15,10 @@ import androidx.core.content.edit
 import androidx.core.content.getSystemService
 import androidx.preference.PreferenceManager
 import com.boswelja.devicemanager.R
-import com.boswelja.devicemanager.common.References.CAPABILITY_PHONE_APP
 import com.boswelja.devicemanager.common.batterysync.BatteryStats
 import com.boswelja.devicemanager.common.batterysync.References.BATTERY_STATUS_PATH
 import com.boswelja.devicemanager.common.batterysync.Utils.updateBatteryStats
+import com.boswelja.devicemanager.common.connection.References.CAPABILITY_PHONE_APP
 import com.boswelja.devicemanager.common.preference.PreferenceKey
 import com.boswelja.devicemanager.phoneconnectionmanager.References.PHONE_NAME_KEY
 import com.google.android.gms.wearable.MessageEvent
